divide: changed -u -> -H, moved some code to Core | explain: Changed formatter | formatter: Added simple number formatter
